refactor. separando a tabela de dados e os modais em componentes separados
